bot/requirements.txt
python-telegram-bot>=21.5
python-dotenv
google-genai
Pillow
fastapi
slowapi
uvicorn[standard]
gunicorn  # Production server with better process management
uvloop  # Fast event loop (2-4x faster than asyncio)
httptools  # Fast HTTP parser
pydantic
python-multipart
SQLAlchemy
alembic
bcrypt>=4.0.0
PyJWT>=2.8.0
# Required for Python 3.13 compatibility
sniffio>=1.3.1
httpx>=0.27.0
anyio>=4.3.0